test_runner refactor for reuse
